+++ ytfetcher/config/__init__.py
@@ -1,7 +1,8 @@
 from .http_config import HTTPConfig
 from .fetch_config import FetchOptions
-from youtube_transcript_api.proxies import ProxyConfig, GenericProxyConfig, WebshareProxyConfig
 from .logging_config import enable_default_config
+from .fetch_config import default_cache_path
+from youtube_transcript_api.proxies import ProxyConfig, GenericProxyConfig, WebshareProxyConfig
 
 __all__ = [
     "HTTPConfig",
@@ -9,5 +10,6 @@ __all__ = [
     "ProxyConfig",
     "GenericProxyConfig",
     "WebshareProxyConfig",
-    "FetchOptions"
+    "FetchOptions",
+    "default_cache_path"
 ]